McAbee Example Data

Leaf physiognomic data of specimens collected from the McAbee Fossil Beds in British Columbia, Canada (Lowe et al. 2018).

Usage

McAbeeExample

Format

McAbeeExample

A data frame with 192 rows and 18 columns:

Site

Stratigraphic layer or locality

Specimen Number

Repository number for individual specimen

Morphotype

Morphotype the specimen belongs to

Measurer comments

Additional notes about the specimen or its measurements

Margin

Whether the margin is toothed (0) or entire (1)

Petiole Width

The width of the petiole at the basalmost point of insertion into the leaf lamina

Blade area

The reconstructed area of the leaf lamina, not including the petiole

Blade perimeter

The length of the perimeter of the leaf lamina, not including the petiole

Feret

The diameter of a circle with the same area as the leaf lamina, not including the petiole

Minimum Feret

The longest line that can be drawn between two points on the perimeter of a selection that is perpendicular to Feret length. Approximates blade width.

Raw blade area

The area of a leaf prepared for tooth measurements that still has its teeth.

Raw blade perimeter

The perimeter of a leaf prepared for tooth measurements that still has its teeth.

Internal raw blade area

The area of a leaf prepared for tooth measurements with teeth digitally removed.

Internal raw blade perimeter

The perimeter of a leaf prepared for tooth measurements with teeth digitally removed.

Length of cut perimeter

The total length of all segments of leaf removed from the leaf blade while removing damage during preparation of the leaf.

no. of primary teeth

The number of primary teeth along the undamaged perimeter

no. of secondary teeth

The number of secondary teeth along the undamaged perimeter

Source

Lowe et al. 2018

References

Lowe, A. J., D. R. Greenwood, C. K. West, J. M. Galloway, M. Sudermann, and T. Reichgelt. 2018. Plant community ecology and climate on an upland volcanic landscape during the Early Eocene Climatic Optimum: McAbee Fossil Beds, British Columbia, Canada. Palaeogeography, Palaeoclimatology, Palaeoecology 511: 433–448.