dilp: Reconstruct Paleoclimate and Paleoecology with Leaf Physiognomy

Use leaf physiognomic methods to reconstruct mean annual temperature (MAT), mean annual precipitation (MAP), and leaf dry mass per area (Ma), along with other useful quantitative leaf traits. Methods in this package described in Lowe et al. (in review).

See also

Author

Maintainer: Matthew J. Butrim mbutrim@uwyo.edu [copyright holder]

Authors:

Alexander J. Lowe

Andrew G. Flynn

Aly Baumgartner

Daniel J. Peppe

Dana L. Royer