Identify outlier specimens

dilp_outliers() is called internally by dilp(). However, it can be used on its own to flag specimens that may have been reported, measured, or prepared incorrectly. dilp_outliers() returns a data frame listing specimens that have unusually high or low values for the four key parameters used in DiLP analyses. This includes whether a specimen is an outlier in the entire dataset, or among other specimens in the same morphotype. If flagged, we suggest looking at the raw measurements and prepped specimen and evaluating if the data is in error or is correct. If in error, the specimen will need to be reprepared and/or remeasured, and the updated datasheet re-read back into R.

Usage

dilp_outliers(specimen_data)

Arguments

specimen_data

Processed specimen level leaf physiognomic data. The structure should match the structure of the output from dilp_processing()

Value

A 4 by X data frame. Each row represents one of the DiLP parameters, and the specimens that are outliers for that parameter.

Examples

# Check for outliers in the provided McAbeeExample dataset. Each
# of these outliers has been manually re-examined and was found acceptable.
dilp_dataset <- dilp_processing(McAbeeExample)
dilp_outliers <- dilp_outliers(dilp_dataset)
#> Warning: Outliers found. Please evaluate $outliers for possible wrong measurements
dilp_outliers
#>    specimen_number         outlier         within morphotype
#> 1     BU-712-1073A perimeter_ratio entire dataset        M28
#> 2      BU-712-1117           tc_ip entire dataset         M8
#> 3      BU-712-1117           tc_ip     morphotype         M8
#> 4      BU-712-1165 perimeter_ratio entire dataset        M28
#> 5     BU-712-1169A           tc_ip entire dataset         M8
#> 6     BU-712-1169A           tc_ip     morphotype         M8
#> 7     BU-712-1176A           tc_ip entire dataset         M8
#> 8     BU-712-1176A           tc_ip     morphotype         M8
#> 9     BU-712-1182A           tc_ip     morphotype         M5
#> 10    BU-712-1182A perimeter_ratio     morphotype         M5
#> 11    BU-712-2105A       leaf_area entire dataset        M47
#> 12     BU-712-2124       leaf_area entire dataset        M94
#> 13    BU-712-2173A       leaf_area entire dataset        M18
#> 14     BU-712-2197 perimeter_ratio     morphotype        M19
#> 15      M-2015-1-1 perimeter_ratio entire dataset        M24
#> 16      M-2015-1-1 perimeter_ratio     morphotype        M24
#> 17    M-2015-1-122 perimeter_ratio     morphotype         M5
#> 18     M-2015-1-17 perimeter_ratio     morphotype        M28
#> 19      M-2015-1-3           tc_ip     morphotype         M5
#> 20      M-2015-1-3 perimeter_ratio     morphotype         M5
#> 21     M-2015-1-40 perimeter_ratio     morphotype         M5
#> 22     M-2015-1-62 perimeter_ratio entire dataset        M28
#> 23     M-2015-1-62 perimeter_ratio     morphotype        M28
#> 24     M-2015-1-69 perimeter_ratio     morphotype         M8
#> 25      M-2015-1-7           tc_ip     morphotype        M19
#> 26     M-2015-2-15           tc_ip     morphotype        M19
#> 27     M-2015-2-84           tc_ip     morphotype        M19